[#27] refactor : 외부와 관련된 클래스모두 external 패키지로 빼기
